Fix conversion_time return type: float→uint32_t to match storage
The getter returned float while the underlying field is uint32_t,
causing silent precision loss after ~4.6h uptime. Use uint32_t
end-to-end and cast to double only at the JSON serialization boundary.

+++ src/temp_simple.c
@@ -70,7 +70,7 @@ float temp_sensor_get_temp(TempSensor *sensor) {
     return sensor->temperature;
 }
 
-float temp_sensor_get_conversion_time(TempSensor *sensor) {
+uint32_t temp_sensor_get_conversion_time(TempSensor *sensor) {
     return sensor->last_conversion_time;
 }
